Fix parsing of the styleable enum/flag attribute values.
Better handling for hexa value > 0x7FFFFFFF

+++ ide_common/src/com/android/ide/common/resources/ValueResourceParser.java
@@ -150,8 +150,10 @@ public final class ValueResourceParser extends DefaultHandler {
                 String value = attributes.getValue(ATTR_VALUE);
 
                 try {
+                    // Integer.decode/parseInt can't deal with hex value > 0x7FFFFFFF so we
+                    // use Long.decode instead.
                     mCurrentDeclareStyleable.addValue(mCurrentAttribute,
-                            name, Integer.decode(value));
+                            name, (int)(long)Long.decode(value));
                 } catch (NumberFormatException e) {
                     // pass, we'll just ignore this value
                 }
